refactor(@dpc-sdp/ripple-tide-grant): :recycle: migrate grant module
