+++ packages/ripple-tide-grant/components/TideGrant/Guidelines.vue
@@ -1,7 +1,3 @@
-<script lang="ts">
-export default { name: 'TideGrantGuidelines' }
-</script>
-
 <template>
   <RplLayoutPageComponent class="tide-grant__guidelines">
     <h2 v-if="title" class="tide-grant__title rpl-type-h2-fixed">
@@ -17,7 +13,7 @@ export default { name: 'TideGrantGuidelines' }
 </template>
 
 <script setup lang="ts">
-import type { TideGrantGuidelineItem } from '../types'
+import type { TideGrantGuidelineItem } from '../../types'
 
 // Redefining TideGrantGuidelines type here since vue can't import child interfaces
 interface Props {
