+++ packages/ripple-tide-grant/package.json
@@ -2,32 +2,10 @@
   "name": "@dpc-sdp/ripple-tide-grant",
   "version": "2.1.0-alpha.52",
   "description": "Ripple mappings and components for Tide Grant Content type",
-  "main": "dist/index.js",
-  "module": "dist/index.js",
-  "type": "module",
-  "exports": {
-    ".": "./dist/index.js",
-    "./component": "./src/index.vue",
-    "./types": "./src/types.ts",
-    "./nuxt": "./src/nuxt.ts"
-  },
+  "main": "nuxt.config.ts",
   "license": "MIT",
   "keywords": [],
-  "scripts": {
-    "clean": "rimraf dist* && rimraf tsconfig.tsbuildinfo",
-    "build": "yarn clean && tsc -p tsconfig.json",
-    "watch": "yarn clean && tsc -p tsconfig.json -w"
-  },
   "dependencies": {
-    "@dpc-sdp/ripple-tide-api": "2.1.0-alpha.52"
-  },
-  "files": [
-    "dist",
-    "src",
-    "!**/*.test.*",
-    "!**/*.json",
-    "CHANGELOG.md",
-    "LICENSE",
-    "README.md"
-  ]
+    "@dpc-sdp/nuxt-ripple": "2.1.0-alpha.52"
+  }
 }
