+++ packages/ripple-tide-grant/server/middleware/grant.ts
@@ -1,0 +1,7 @@
+import { defineEventHandler } from 'h3'
+import { registerTideContentType } from '@dpc-sdp/ripple-tide-api'
+import tideGrantModule from '../../mapping'
+
+export default defineEventHandler(async (event) => {
+  registerTideContentType(event, 'grant', tideGrantModule)
+})
